Move the routes to a sub-directory. Rename the model directory.
