+++ src/App.scss
@@ -1,4 +1,9 @@
+// Personal library styles
 @use "@tahoni/tahoni-lib-react/dist/index.css" as *;
+// Project styles
 @use "assets/stylesheets/style" as *;
+// External library styles
 @use "vendors/bootstrap/stylesheets/style" as *;
+
+// Package styles
 @use "sweetalert2/dist/sweetalert2.css" as *;
