+++ src/components/Map/VenuePins.tsx
@@ -1,7 +1,7 @@
-import React, {ReactElement} from "react";
+import React, { ReactElement } from "react";
 import VenuePin from "./VenuePin.tsx";
-import {VenueType} from "../../model/Venue.ts";
-import {generateMapVenueKey} from "../../utils/MapUtils.ts";
+import { VenueType } from "../../models/Venue.ts";
+import { generateMapVenueKey } from "../../utils/MapUtils.ts";
 
 export interface VenuePinsProps {
   venues: VenueType[];
@@ -12,7 +12,9 @@ const VenuePins = React.memo((props: VenuePinsProps): ReactElement => {
     return <></>;
   }
 
-  const venues: VenueType[] = props.venues.filter((venue) => venue).map((venue) => venue);
+  const venues: VenueType[] = props.venues
+    .filter((venue) => venue)
+    .map((venue) => venue);
 
   return (
     <>
