+++ src/config/Routes/AppRoutes.tsx
@@ -3,20 +3,17 @@ import { Route, Routes } from "react-router";
 import {
   leftShooter,
   rightShooter,
-} from "../constants/images/LayoutImageConstants";
-import { PageAlias } from "../model/PageAlias";
-import { routes } from "../helpers/RouteHelpers";
-import { Layout } from "../layout";
+} from "../../constants/images/LayoutImageConstants.ts";
+import { PageAlias } from "../../models/PageAlias.ts";
+import { routes } from "../../helpers/RouteHelpers.tsx";
+import { Layout } from "../../layout";
 
 const AppRoutes = (): ReactElement => {
   return (
     <Routes>
       <Route
         element={
-          <Layout
-            leftSideImage={leftShooter}
-            rightSideImage={rightShooter}
-          />
+          <Layout leftSideImage={leftShooter} rightSideImage={rightShooter} />
         }
       >
         {routes.map((route: PageAlias, index) => (
