+++ src/helpers/MenuHelpers.tsx
@@ -1,4 +1,4 @@
-import { PageMapping } from "../model/PageMapping.ts";
+import { PageMapping } from "../models/PageMapping.ts";
 import {
   aboutUs,
   events,
@@ -6,7 +6,7 @@ import {
   home,
   links,
   members,
-} from "../config/RouteAliases.tsx";
+} from "../config/Routes/RouteAliases.tsx";
 
 export const menuItems: PageMapping[] = [
   home,
