+++ src/models/Email.ts
