+++ src/models/EmailAttachment.ts
