+++ src/models/EmailContent.ts
