+++ src/models/EmailMessage.ts
