+++ src/models/VenueEvent.ts
