+++ src/vendors/bootstrap/stylesheets/style.scss
@@ -1,14 +1,14 @@
-// Required
+// Required Bootstrap functions
 @import "../../../../node_modules/bootstrap/scss/functions";
 
 // Default variable overrides
 @import "custom";
 
-// Required
+// Required Bootstrap components
 @import "../../../../node_modules/bootstrap/scss/variables";
 @import "../../../../node_modules/bootstrap/scss/variables-dark";
 @import "../../../../node_modules/bootstrap/scss/maps";
 @import "../../../../node_modules/bootstrap/scss/mixins";
 @import "../../../../node_modules/bootstrap/scss/root";
 
-// Optional Bootstrap components here
+// Optional Bootstrap components
